+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - PR #402 Install dependencies via rapids-build-env
 - PR #405 Move doc customization scripts to Jenkins
 - PR #414 Add element-wise access for device_uvector
+- PR #421 Capture thread id in logging and improve logger testing
 
 ## Bug Fixes